Anand - Reverted add guardian
